Move judgeRemaining method to trait so we are sure $this->em is defined everywhere
